create custom hook that handles changing the scrollbar width css variable
